requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
ultralytics==8.0.208
openai==1.3.5
arabic-reshaper==3.0.0
python-bidi==0.4.2
langdetect==1.0.9
elevenlabs==0.2.26
word2number==1.1
PyMuPDF==1.23.7
python-dotenv==1.0.0
pydantic==2.5.2
typing-extensions==4.8.0